Handle multiple Accept headers
According to RFC2616, a header may be repeated any time the header's value is a comma separated list of values, in which case multiple headers should be treated as multiple comma-separated values.  We weren't handling that before; I would much rather that we handled that than rely on the client code to know about this part of RFC2616.